Move KnockbackSettings to knockback package
