+++ src/main/java/io/github/togar2/pvp/feature/knockback/KnockbackSettings.java
@@ -1,4 +1,4 @@
-package io.github.togar2.pvp.legacy;
+package io.github.togar2.pvp.feature.knockback;
 
 import net.minestom.server.ServerFlag;
