+++ src/main/java/io/github/togar2/pvp/feature/knockback/VanillaKnockbackFeature.java
@@ -4,7 +4,6 @@ import io.github.togar2.pvp.events.EntityKnockbackEvent;
 import io.github.togar2.pvp.feature.FeatureType;
 import io.github.togar2.pvp.feature.config.DefinedFeature;
 import io.github.togar2.pvp.feature.config.FeatureConfiguration;
-import io.github.togar2.pvp.legacy.KnockbackSettings;
 import io.github.togar2.pvp.player.CombatPlayer;
 import io.github.togar2.pvp.utils.CombatVersion;
 import net.minestom.server.coordinate.Vec;
